+++ lib/src/native/build_all.sh
@@ -31,6 +31,6 @@ if [ ! -f "$FILE" ]; then
     exit 1
 fi
 
-gcc -I/usr/local/include/periphery/  -Wall -fPIC -O3 -shared gpio.c serial.c i2c.c led.c pwm.c version.c spi.c -lperiphery  -lpthread -o dart_periphery_${BITNESS}.1.0.0.so
-gcc -I/usr/local/include/periphery/  -Wall -fPIC -O3 -shared gpio.c serial.c i2c.c led.c pwm.c version.c spi.c  libperiphery_${BITNESS}.a  -lpthread -o dart_periphery_static_${BITNESS}.1.0.0.so
+gcc -I/usr/local/include/periphery/  -Wall -fPIC -O3 -shared gpio.c serial.c i2c.c led.c pwm.c version.c spi.c mmio.c -lperiphery  -lpthread -o dart_periphery_${BITNESS}.1.0.0.so
+gcc -I/usr/local/include/periphery/  -Wall -fPIC -O3 -shared gpio.c serial.c i2c.c led.c pwm.c version.c spi.c mmio.c libperiphery_${BITNESS}.a  -lpthread -o dart_periphery_static_${BITNESS}.1.0.0.so